+++ src/main.rs
@@ -4,562 +4,10 @@ use std::env;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Mutex;
 use std::collections::HashMap;
-use bio::io::fasta;
 use std::sync::Arc;
 
-
-// We are still not filtering out based on TLEN (python version does this)
-// We are not skipping the whole read when INDELS (python version does this)
-// Still needs to check if it's handling ONT data properly.
-
-
-#[derive(Debug, Clone, Hash, Eq, PartialEq)]
-struct MismatchKey {
-    mismatch_type: String,  // e.g., "A>G", "C>T"
-    read_position: usize,   // Position in the read (0-based)
-    read_num: u8,           // 1 or 2 for paired-end reads
-}
-
-// Track inconsistencies between read1 and read2 in overlap regions
-#[derive(Debug, Clone, Hash, Eq, PartialEq)]
-struct InconsistencyKey {
-    discordance_type: String,  // e.g., "R1:A_R2:G"
-    read1_position: usize,
-    read2_position: usize,
-}
-
-// Store read information for overlap detection
-#[derive(Debug, Clone)]
-struct ReadInfo {
-    qname: Vec<u8>,
-    tid: i32,
-    pos: i64,
-    seq_len: usize,
-    cigar: Vec<u8>,  // Store cigar as bytes for later use
-    is_reverse: bool,
-    read_num: u8,
-}
-
-// Store reference sequences in memory
-type ReferenceGenome = HashMap<String, Vec<u8>>;
-
-fn load_reference_genome(fasta_path: &str) -> ReferenceGenome {
-    eprintln!("Loading reference genome from: {}", fasta_path);
-    let reader = fasta::Reader::from_file(fasta_path)
-        .expect("Failed to open reference FASTA file");
-    
-    let mut genome: ReferenceGenome = HashMap::new();
-    for result in reader.records() {
-        let record = result.expect("Failed to read FASTA record");
-        let chr_name = record.id().to_string();
-        let sequence = record.seq().to_vec(); //Vec<u8> = byte, not UTF-8 char (overhead) 
-        genome.insert(chr_name, sequence);
-    }
-    
-    eprintln!("Loaded {} chromosome(s)", genome.len());
-    genome
-}
-
-fn complement(base: char) -> char {
-    match base {
-        'A' => 'T',
-        'T' => 'A',
-        'C' => 'G',
-        'G' => 'C',
-        _ => base,
-    }
-}
-
-fn base_to_char(byte: u8) -> Option<char> {
-    match byte {
-        b'A' => Some('A'),
-        b'C' => Some('C'),
-        b'G' => Some('G'),
-        b'T' => Some('T'),
-        _ => None,
-    }
-}
-
-fn parse_md_tag(md_string: &str) -> (Vec<(usize, char)>, Vec<(usize, usize)>) {
-    // Parse MD tag and return:
-    // 1. (position_offset, ref_base) for mismatches
-    // 2. (start_pos, length) for matches
-    let mut mismatches = Vec::new();
-    let mut matches = Vec::new();
-    let mut pos = 0;
-    let mut num_str = String::new();
-    
-    for ch in md_string.chars() {
-        if ch.is_numeric() {
-            num_str.push(ch);
-        } else if ch == '^' {
-            // Deletion - skip the deleted bases
-            if !num_str.is_empty() {
-                let match_len = num_str.parse::<usize>().unwrap_or(0);
-                if match_len > 0 {
-                    matches.push((pos, match_len));
-                    pos += match_len;
-                }
-                num_str.clear();
-            }
-        } else if ch.is_alphabetic() {
-            // This is a reference base at a mismatch position
-            if !num_str.is_empty() {
-                let match_len = num_str.parse::<usize>().unwrap_or(0);
-                if match_len > 0 {
-                    matches.push((pos, match_len));
-                    pos += match_len;
-                }
-                num_str.clear();
-            }
-            mismatches.push((pos, ch));
-            pos += 1;
-        }
-    }
-    
-    // Handle trailing matches
-    if !num_str.is_empty() {
-        let match_len = num_str.parse::<usize>().unwrap_or(0);
-        if match_len > 0 {
-            matches.push((pos, match_len));
-        }
-    }
-    
-    (mismatches, matches)
-}
-
-fn compute_read_len_mode_from_sample_bam(bam_path: &str, sample_size: usize) -> usize {
-    let mut bam = Reader::from_path(bam_path)
-        .expect("Failed to open BAM file for read length sampling");
-    
-    let mut length_counts: HashMap<usize, usize> = HashMap::new();
-    
-    for (_i, result) in bam.records().enumerate().take(sample_size) {
-        if let Ok(record) = result {
-            let read_len = record.seq().len();
-            *length_counts.entry(read_len).or_insert(0) += 1;
-        }
-    }
-    
-    // Find mode of read lengths
-    let mode_length = length_counts.iter()
-                                    .max_by_key(|&(_, count)| count)
-                                    .map(|(&key, _)| key)
-                                    .unwrap_or(0);
-
-    eprintln!("Computed mode read length from sample: {}", mode_length);
-    mode_length
-}
-
-fn adjust_methylation_base(
-    read_base: char,
-    ref_base: char,
-    read_num: u8,
-    is_methylation: bool,
-    cpg_only: bool,
-    ref_seq: &[u8],
-    genome_pos: usize
-) -> char {
-    if !is_methylation {
-        return read_base;
-    }
-    
-    if cpg_only {
-        // Only convert in CpG context
-        match (read_num, ref_base, read_base) {
-            // Read 1: C in ref, next base is G, T in read -> treat as C (unmethylated CpG)
-            (1, 'C', 'T') => {
-                if genome_pos + 1 < ref_seq.len() {
-                    let next_base = ref_seq[genome_pos + 1];
-                    if next_base == b'G' || next_base == b'g' {
-                        'C'
-                    } else {
-                        read_base
-                    }
-                } else {
-                    read_base
-                }
-            }
-            // Read 2: G in ref, next base is C, A in read -> treat as G (unmethylated CpG on reverse)
-            (2, 'G', 'A') => {
-                if genome_pos + 1 < ref_seq.len() {
-                    let next_base = ref_seq[genome_pos + 1];
-                    if next_base == b'C' || next_base == b'c' {
-                        'G'
-                    } else {
-                        read_base
-                    }
-                } else {
-                    read_base
-                }
-            }
-            _ => read_base,
-        }
-    } else {
-        // Convert all C>T and G>A
-        match (read_num, ref_base, read_base) {
-            // Read 1: C in ref, T in read -> treat as C (unmethylated C)
-            (1, 'C', 'T') => 'C',
-            // Read 2: G in ref, A in read -> treat as G (unmethylated C on reverse strand)
-            (2, 'G', 'A') => 'G',
-            _ => read_base,
-        }
-    }
-}
-
-fn create_mismatch_key(
-    read_base: char,
-    ref_base: char,
-    r_pos: usize,
-    seq_len: usize,
-    is_reverse: bool,
-    read_num: u8,
-    is_methylation: bool,
-    cpg_only: bool,
-    ref_seq: &[u8],
-    genome_pos: usize,
-    mode_len: usize
-) -> MismatchKey {
-    // Apply reverse complement if read is mapped to reverse strand
-    let strand_adjusted_read_base = if is_reverse { complement(read_base) } else { read_base };
-    let strand_adjusted_ref_base = if is_reverse { complement(ref_base) } else { ref_base };
-
-    // Apply methylation-aware base conversion if enabled
-    let meth_and_strand_adjusted_read_base = adjust_methylation_base(
-        strand_adjusted_read_base, 
-        strand_adjusted_ref_base, 
-        read_num, 
-        is_methylation, 
-        cpg_only, 
-        ref_seq, 
-        genome_pos
-    );
-    let adjusted_r_pos = if is_reverse {seq_len - r_pos -1} else {r_pos};
-    let mode_adjusted_r_pos = correct_read_len_with_mode(adjusted_r_pos, seq_len, mode_len);
-
-    MismatchKey {
-        mismatch_type: format!("{}>{}", strand_adjusted_ref_base, meth_and_strand_adjusted_read_base),
-        read_position: mode_adjusted_r_pos,
-        read_num,
-    }
-}
-
-fn compare_and_count(
-    seq: &rust_htslib::bam::record::Seq,
-    qual: &[u8],
-    ref_seq: &[u8],
-    r_pos: usize,
-    genome_pos: usize,
-    is_reverse: bool,
-    read_num: u8,
-    min_base_quality: u8,
-    is_methylation: bool,
-    cpg_only: bool,
-    local_counts: &mut HashMap<MismatchKey, usize>,
-    mode_len: usize
-) {
-    let seq_len = seq.len();
-    let ref_len = ref_seq.len();
-    
-    if r_pos >= seq_len || genome_pos >= ref_len { return; }
-    
-    // Check base quality score (skip if below threshold)
-    if r_pos >= qual.len() || qual[r_pos] < min_base_quality { return; }
-    
-    let Some(read_base) = base_to_char(seq[r_pos]) else { return; };    
-    let Some(ref_base) = base_to_char(ref_seq[genome_pos]) else { return; };
-    
-    // Count both matches and mismatches
-    let key = create_mismatch_key(read_base, ref_base, r_pos, seq_len, is_reverse, read_num, is_methylation, cpg_only, ref_seq, genome_pos, mode_len);
-    *local_counts.entry(key).or_insert(0) += 1;
-}
-
-fn correct_read_len_with_mode(
-    read_pos: usize,
-    seq_len: usize,
-    mode_len: usize
-) -> usize {
-    if mode_len > 0 && read_pos > seq_len/2 {  
-        return read_pos + (mode_len - seq_len);
-    } else {
-        return read_pos;
-    }
-}
-
-// Calculate genomic end position from CIGAR
-fn calculate_end_pos(start_pos: i64, cigar: &rust_htslib::bam::record::CigarStringView) -> i64 {
-    use rust_htslib::bam::record::Cigar::*;
-    let mut end = start_pos;
-    for op in cigar.iter() {
-        match op {
-            Match(len) | Equal(len) | Diff(len) | Del(len) | RefSkip(len) => {
-                end += *len as i64;
-            }
-            _ => {}
-        }
-    }
-    end
-}
-
-// Check if two reads overlap genomically
-fn get_overlap_region(read1: &ReadInfo, read2: &ReadInfo, record1: &Record, record2: &Record) -> Option<(i64, i64)> {
-    if read1.tid != read2.tid {
-        return None;
-    }
-    
-    let end1 = calculate_end_pos(read1.pos, &record1.cigar());
-    let end2 = calculate_end_pos(read2.pos, &record2.cigar());
-    
-    let overlap_start = read1.pos.max(read2.pos);
-    let overlap_end = end1.min(end2);
-    
-    if overlap_start < overlap_end {
-        Some((overlap_start, overlap_end))
-    } else {
-        None
-    }
-}
-
-// Process overlap region between two reads
-// Process both reads - MismatchKey has read_num so they're counted separately
-// Also detect inconsistencies between read1 and read2
-fn process_overlap_region(
-    record1: &Record,
-    record2: &Record,
-    overlap_start: i64,
-    overlap_end: i64,
-    local_counts: &mut HashMap<MismatchKey, usize>,
-    inconsistency_counts: &mut HashMap<InconsistencyKey, usize>,
-    reference: &ReferenceGenome,
-    tid_to_name: &HashMap<i32, String>,
-    min_base_quality: u8,
-    is_methylation: bool,
-    cpg_only: bool,
-    mode_len: usize,
-) {
-    // Build genome_pos -> (read_pos, base, qual) mapping for both reads
-    let mut read1_map: HashMap<usize, (usize, u8, u8)> = HashMap::new();
-    let mut read2_map: HashMap<usize, (usize, u8, u8)> = HashMap::new();
-    
-    for (record, map) in [(record1, &mut read1_map), (record2, &mut read2_map)] {
-        if record.is_unmapped() || record.mapq() <= min_map_quality {
-            continue;
-        }
-        
-        let ref_start = record.pos() as usize;
-        let seq = record.seq();
-        let qual = record.qual();
-        let cigar = record.cigar();
-        
-        let mut read_pos = 0;
-        let mut ref_pos = ref_start;
-        
-        use rust_htslib::bam::record::Cigar::*;
-        for cigar_op in cigar.iter() {
-            match cigar_op {
-                Match(len) | Equal(len) | Diff(len) => {
-                    for i in 0..*len {
-                        let genome_pos = ref_pos + i as usize;
-                        if genome_pos >= overlap_start as usize && genome_pos < overlap_end as usize {
-                            let r_pos = read_pos + i as usize;
-                            if r_pos < seq.len() && r_pos < qual.len() {
-                                map.insert(genome_pos, (r_pos, seq[r_pos], qual[r_pos]));
-                            }
-                        }
-                    }
-                    read_pos += *len as usize;
-                    ref_pos += *len as usize;
-                }
-                Ins(len) => { read_pos += *len as usize; }
-                Del(len) | RefSkip(len) => { ref_pos += *len as usize; }
-                _ => {}
-            }
-        }
-    }
-    
-    // Compare bases at overlapping genomic positions
-    for (genome_pos, (r1_pos, r1_base, r1_qual)) in read1_map.iter() {
-        if let Some((r2_pos, r2_base, r2_qual)) = read2_map.get(genome_pos) {
-            // Both reads cover this position - check for inconsistency
-            if r1_qual >= &min_base_quality && r2_qual >= &min_base_quality {
-                if r1_base != r2_base {
-                    // Bases disagree - record inconsistency
-                    let r1_char = base_to_char(*r1_base).unwrap_or('N');
-                    let r2_char = base_to_char(*r2_base).unwrap_or('N');
-                    
-                    let key = InconsistencyKey {
-                        discordance_type: format!("R1:{}_R2:{}", r1_char, r2_char),
-                        read1_position: *r1_pos,
-                        read2_position: *r2_pos,
-                    };
-                    *inconsistency_counts.entry(key).or_insert(0) += 1;
-                }
-            }
-        }
-    }
-    
-    // Process mismatch counts for both reads
-    for record in [record1, record2] {
-        if record.is_unmapped() || record.mapq() <= min_map_quality {
-            continue;
-        }
-        
-        let tid = record.tid();
-        let Some(chr_name) = tid_to_name.get(&tid) else { continue };
-        let Some(ref_seq) = reference.get(chr_name.as_str()) else { continue };
-        let ref_start = record.pos() as usize;
-        let seq = record.seq();
-        let qual = record.qual();
-        let cigar = record.cigar();
-        let read_num = if record.is_first_in_template() { 1 } else if record.is_last_in_template() { 2 } else { 1 };
-        
-        let mut read_pos = 0;
-        let mut ref_pos = ref_start;
-        
-        use rust_htslib::bam::record::Cigar::*;
-        for cigar_op in cigar.iter() {
-            match cigar_op {
-                Match(len) | Equal(len) | Diff(len) => {
-                    for i in 0..*len {
-                        let genome_pos = ref_pos + i as usize;
-                        
-                        // Only count if in overlap region
-                        if genome_pos >= overlap_start as usize && genome_pos < overlap_end as usize {
-                            let r_pos = read_pos + i as usize;
-                            compare_and_count(
-                                &seq, &qual, ref_seq, r_pos, genome_pos,
-                                record.is_reverse(), read_num, min_base_quality,
-                                is_methylation, cpg_only, local_counts, mode_len
-                            );
-                        }
-                    }
-                    read_pos += *len as usize;
-                    ref_pos += *len as usize;
-                }
-                Ins(len) => { read_pos += *len as usize; }
-                Del(len) | RefSkip(len) => { ref_pos += *len as usize; }
-                _ => {}
-            }
-        }
-    }
-}
-
-// Calculate genomic end position from CIGAR
-fn process_record(
-    record: &Record, 
-    local_counts: &mut HashMap<MismatchKey, usize>, 
-    reference: &ReferenceGenome, 
-    tid_to_name: &HashMap<i32, String>,
-    softclip_threshold: f64,
-    min_base_quality: u8,
-    is_methylation: bool,
-    cpg_only: bool,
-    mode_len: usize
-) {
-    
-    // Skip read
-    if record.is_unmapped() 
-        || record.is_secondary() 
-        || record.is_supplementary() 
-        || record.mapq() <= min_map_quality
-    { return; }
-    
-    // Get chromosome name (target_id - defined from bam header) and other read info
-    let tid = record.tid();
-    if tid < 0 { return; } // unmapped    
-    let Some(chr_name) = tid_to_name.get(&tid) else { return }; // This directly binds chr_name (extract from Option or return)
-    let Some(ref_seq) = reference.get(chr_name.as_str()) else { return; };
-    let ref_start = record.pos() as usize; // (0-based)
-    let seq = record.seq();
-    let qual = record.qual();
-    let cigar = record.cigar();
-    let read_num = if record.is_first_in_template() { 1 } else if record.is_last_in_template() { 2 } else { 1 };
-    
-    // Walk through CIGAR and compare read to reference
-    let mut read_pos = 0;
-    let mut ref_pos = ref_start;
-    
-    for cigar_op in cigar.iter() {
-        use rust_htslib::bam::record::Cigar::*;
-        match cigar_op {
-            Match(len) | Equal(len) | Diff(len) => {
-                // Compare read and reference bases
-                for i in 0..*len {
-                    let r_pos = read_pos + i as usize; //if mode_len > 0 { mode_len - read_pos + i as usize } else { read_pos + i as usize };
-                    let genome_pos = ref_pos + i as usize;
-                    compare_and_count(&seq, &qual, ref_seq, r_pos, genome_pos, record.is_reverse(), read_num, min_base_quality, is_methylation, cpg_only, local_counts, mode_len);
-                }
-                read_pos += *len as usize;
-                ref_pos += *len as usize;
-            }
-            SoftClip(len) => {
-                // Soft-clipped bases - only count mismatches if at least 66% of bases match reference
-                let mut total_bases = 0;
-                let mut matching_bases = 0;
-                let mut temp_mismatch_keys = Vec::<MismatchKey>::new();
-                let seq_len = seq.len();
-                
-                for i in 0..*len {
-                    let r_pos = read_pos + i as usize;
-                    let genome_pos = if read_pos == 0 { // beginning of read
-                        if ref_pos >= (*len - i) as usize {
-                            ref_pos - (*len - i) as usize
-                        } else {
-                            continue;
-                        }
-                    } else { // end of read
-                        ref_pos + i as usize
-                    };
-                    
-                    if r_pos >= seq_len || genome_pos >= ref_seq.len() {
-                        continue;
-                    }
-                    
-                    // Skip bases with low quality
-                    if r_pos >= qual.len() || qual[r_pos] < min_base_quality {
-                        continue;
-                    }
-                    
-                    let read_base = match seq[r_pos] {
-                        b'A' => 'A',
-                        b'C' => 'C',
-                        b'G' => 'G',
-                        b'T' => 'T',
-                        _ => continue,
-                    };
-                    
-                    let ref_base = match ref_seq[genome_pos] {
-                        b'A' | b'a' => 'A',
-                        b'C' | b'c' => 'C',
-                        b'G' | b'g' => 'G',
-                        b'T' | b't' => 'T',
-                        _ => continue,
-                    };
-                    
-                    total_bases += 1;
-                    if read_base == ref_base {
-                        matching_bases += 1;
-                    } else {
-                        // Create MismatchKey using helper function
-                        let key = create_mismatch_key(read_base, ref_base, r_pos, seq_len, record.is_reverse(), read_num, is_methylation, cpg_only, ref_seq, genome_pos, mode_len);
-                        temp_mismatch_keys.push(key);
-                    }
-                }
-                
-                // Only add mismatches to counts if at least 66% (default) match
-                if total_bases > 0 && (matching_bases as f64 / total_bases as f64) >= softclip_threshold {
-                    for key in temp_mismatch_keys {
-                        *local_counts.entry(key).or_insert(0) += 1;
-                    }
-                }
-                read_pos += *len as usize;
-            }
-            Ins(len) => { read_pos += *len as usize; }
-            Del(len) | RefSkip(len) => { ref_pos += *len as usize; }
-            HardClip(_) | Pad(_) => {}
-        }
-    }
-}
+// Import from library
+use rustmanian_mismatch::*;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -581,7 +29,7 @@ fn main() {
     let mut is_methylation: bool = false; // Methylation-aware mode (for bisulfite/EM-seq)
     let mut cpg_only: bool = false; // Only apply methylation conversion in CpG context
     let mut mode_len = 0;
-    let mut insert_position_mode = false; // Not implemented yet (Not sure about it yet...)
+    let mut _insert_position_mode = false; // Not implemented yet (Not sure about it yet...)
 
     let mut i = 3;
     while i < args.len() {
@@ -643,7 +91,7 @@ fn main() {
             }
             "--insert-position-mode" => {
                 eprintln!("Using --insert-position-mode. Not read position");
-                insert_position_mode = true;
+                _insert_position_mode = true;
                 i += 1;
             }
             _ => {
@@ -746,7 +194,7 @@ fn main() {
                     // Process normally
                     process_record(
                         &record, &mut region_counts, &ref_clone, &tid_clone, softclip_threshold, 
-                        min_base_quality, is_methylation, cpg_only, mode_len
+                        min_base_quality, is_methylation, cpg_only, mode_len, min_map_quality
                     );
                     local_count += 1;
                     
@@ -754,13 +202,8 @@ fn main() {
                     if record.is_paired() && !record.is_unmapped() && !record.is_mate_unmapped() {
                         let qname = record.qname().to_vec();
                         let read_info = ReadInfo {
-                            qname: qname.clone(),
                             tid: record.tid(),
                             pos: record.pos(),
-                            seq_len: record.seq().len(),
-                            cigar: Vec::new(),
-                            is_reverse: record.is_reverse(),
-                            read_num: if record.is_first_in_template() { 1 } else if record.is_last_in_template() { 2 } else { 1 },
                         };
                         
                         // Check if mate already cached
@@ -773,7 +216,7 @@ fn main() {
                                     &record, &mate_record, overlap_start, overlap_end,
                                     &mut overlap_region_counts, &mut inconsistency_region_counts,
                                     &ref_clone, &tid_clone,
-                                    min_base_quality, is_methylation, cpg_only, mode_len
+                                    min_base_quality, is_methylation, cpg_only, mode_len, min_map_quality
                                 );
                             }
                         } else {
